Improved notification profile models :card_file_box:
Removed a lot of the old migrations and added a couple migrations that delete the NotificationProfile model and then re-add it together with the new models (Filter, TimeSlotGroup, TimeSlot). The NotificationProfile model was deleted to not have to provide a default value for the new required foreign key fields: `time_slot_group` and `filters`.

Also:
 * Changed `start` and `end` of TimeSlot from `DateTimeField` to just `TimeField`, which easier allows the frontend to post time slots with only time and no date
 * Moved `filters` field from TimeSlotGroup to NotificationProfile
 * Renamed `Filter.filter` to `Filter.query`
 * Added `active` field to NotificationProfile