Changed from UTC time slots to local timestamp :bug:
The previous logic converted time slots' times to UTC and then checked if an alert's timestamp was between the time slot's start and end; this did not work as intended in some edge cases, like with a `TimeSlot(start=00:00, end=23:59)` which - when converted - would become `TimeSlot(utc_start=22:00, utc_end=21:59)`, which doesn't match any timestamps.

This commit changes this logic to compare a time slot's times with timestamps converted to local time.

Also removed filtering historic alerts on time slots, as it's something one is rarely interested in.