Replace existing migrations
As this commit replaces all existing migrations entirely, it is easiest
to drop the database and start afresh.

Each table name has been prefixed with "aas_", and "notificationprofile"
has been shortened to "notoprofile" just in the sql since there are so
many max length restrictions on auto-generated names we might run into.