Rename TimeSlot to Timeslot
The reasons for this are that it makes the naming more consistent between the frontend and backend, and we felt that it looked more natural.

The original name was chosen based on that the majority of dictionaries seemed to mention "time slot" as the preferable way to spell it; "timeslot" also appears in several dictionaries, though.