Fix AlertSourceList returning None serializer
This should not have been a problem, but when running the server through Docker, it turns out that there is somehow sent both one `GET` and one `POST` request when visiting `/api/v1/alerts/sources/` in Chrome (posting a request through e.g. Postman does not have this effect). Furthermore, the `POST` one does not call the overridden `create()` method of `AlertSourceList` for some reason, which would have caused `get_serializer_class()` to not get called.