Implement real-time incidents API using WebSockets
This patch introduces a WebSockets based real-time
active incidents API. It works by making clients
subscribe to active incidents and uses Django signals to
catch changes and notify all the clients.

For now it uses the synchronous websockets JSON consumer,
but it should probably be changed to use the async version.