Use explicit paths instead of reverse paths
This makes it alot easier to see which endpoints are
actually being tested. The comments that are removed in this commit
were added by me previously because they were absolutely
necessary for me to keep track of what the result of the
reverse function actually was. I also had to spend a
considerable amount of time when starting to write the tests
to figure out how the reverse function even worked.
So explicit pathing also makes it easier for other people
to expand/change the tests.