Install editable version of argus-server
The /api/ endpoint returns the version number of the server. In a
development install, this is only available through the distribution
information for `argus-server` - yet `argus-server` is not actually
installed in the Docker image, it is only mounted as a source code
directory. Thus, the /api/ endpoint crashes.

This mitigates this by ensuring an editable (symlinked) installation of
the source code is made every time the container starts.