Output test reports in tox aware output directory
This sets up test-reports as the default directory for the CI pipeline
to pick up test report artifacts. Additionally, if we appear to be
running in a tox environment, we separate the report files in
subdirectories based on tox environment name. This way, when the same
test is run in multiple environments, we can easily detect in which ones
the failures were in.