Add qcluster as a separate Docker Compose service
This sets up the API web server and the Django Q2 qcluster command as
separate Docker Compose services, based on the same image and config.

The only difference between the two containers is which command they
run.  The image defaults to the API entrypoint, while docker-compose.yml
will override the command of the qcluster service.

In an attempt to make it DRY, the two nearly-identical services are
templated out using YAML anchors and X properties, as suggested by the
Docker Compose docs.