Link two first columns in Incident admin
By default, the first displayed field is linked to the details page. The
first field we show is `source_incident_id` which is optional. When it
is not set there will be no link. Therefore, also link up the second
field (`start_time`) so that we can always easily get to the details
page.