Move navbar notification links to user menu
Moving these left only one link behind, the redundant link to
"incidents" which is also reachable via clicking the logo. This did not
look good, so that link was also removed, emptying the navbar save for
the logo and user icon.

A commented out example is left behind to show what customized links
need to look like.