Use text type for timeinput
Makes it so you have to type in HH:MM format with little help
other than placeholder. type=time is nicer, but depending on the
local language of your computer it might show 24h or AM/PM.
We want to avoid AM/PM at pretty much all costs so we use this
basic field for selecting time with goal of making
something nicer later that will be reused other places.