Make basic inputs in forms use a universal design
Uses a new generic template for universal basic input field and applies it to
all modals, the timeslots page and the destination page.

Not covered: profiles page (to come). Also does not apply to inputs in the
incident filter box since the look isn't stable and they are more of a control
interface than a form.

This is not the final method to achieve consistency. We need to send in the
forms into the modals, which will simplify `input_field.html`, and Django 5.2
adds some features that should do the overriding more elegant.