Make tags containing HTTP URL clickable on incident details page
If the value of a tag is a valid url, turn it into a link.

Rather than using arbitrary substring logic in template code, this defers to
Python/Django library code to actually validate the full tag value as a valid
URL before hyperlinking it via a new URL validation template filter,
`is_valid_url`.

---------

Co-authored-by: Johanna England <johanna.england@sikt.no>
Co-authored-by: Morten Brekkevold <morten.brekkevold@sikt.no>