Explicitly install SPA requirements in Docker image
The Docker image as currently defined would not start because it is
missing the `channels` Python package.  The Docker image is still built
to work with the SPA front-end, so it uses the websockets/uvicorn
setup.

However, since Argus 1.33, the procedure for setting up Argus with
the SPA front-end changed, while this Dockerfile did not.  The
necessary requirements for SPA front-end compatibility are now defined
by the optional `spa` requirements in `pyproject.toml`.