Make a model-independent ComplexFilterWrapper
The NotificationProfile-specific ComplexFilterWrapper has been renamed to
NotificationProfileFilterWrapper and moved to the notificationprofile app. In
the process some missing unit tests only useful for matching up
NotificationProfiles were added.
                                                                                
The new PreciscionFilterWrapper only cares about matching combined filters, and
not what the filters are to be used for.
                                                                                
ComplexFallbackFilterWrapper is no longer needed, it was only useful for
NotificationProfiles.